Refactor mobile
